requirements.txt
# Core requirements
requests==2.31.0
tqdm==4.66.1
scikit-learn==1.3.2
Pillow==10.0.1

# PyTorch ecosystem (choose appropriate version for your CUDA)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Additional utilities
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Logging and experiment tracking
tensorboard==2.13.0